Fix regression at "Refresh" feature partially
The application is supposed to stay inside the same document (after
some navigation was done) and to keep the scroll position. This got
broken years ago with commit 41352981d866cd0e130595114963f70785e13936

Now the application stays in the same document again. It was given up
to fix the scroll position problem. The logic handling will be
completely reimplemented in "next" major version.