Open menu at clicking on unblock button
Click on the banner itself doesn't do anything anymore. Also, the main
menu was changed that "Unblock All External Content" is now a submenu.
This justifies the major version change.

Because it is not possible to test popup menus in Playwright, the UI
test for unblocking via banner click is removed without replacement.