First approach for content blocking dialog
It is planned to present a list of all blocked URLs to the user. These
URLs shall all be selected by default, so the user has only to press
the "OK" button (or press Enter) to unblock all. In turn, it shall be
possible to unselect URLs, so they stay blocked.